Add files via upload
Added Test page which uses useFormik. There are only 3 input fields, Name', 'Address 1', and 'Reason for Extension'. I used what I learned from researching for the interview and implemented it as another solution to the reason for extension box. All 3 fields are required, have forMik.touched, and the 'Reason for Extension' shows an error if there are <= 10 characters.